refactor(domain): reorganize imports and move CellRange and FlattenDoubleVec to infrastructure module
