+++ app/src/application/price/token_prices.rs
@@ -1,7 +1,7 @@
 use crate::domain::routine::{Routine, RoutineError};
-use crate::domain::sheets::flatten_double_vec::FlattenDoubleVec;
 use crate::domain::sheets::ranges;
 use crate::infrastructure::price::price::get_token_prices;
+use crate::infrastructure::sheets::flatten_double_vec::FlattenDoubleVec;
 use crate::infrastructure::sheets::spreadsheet_manager::SpreadsheetManager;
 use crate::infrastructure::sheets::spreadsheet_read::SpreadsheetRead;
 use crate::infrastructure::sheets::spreadsheet_write::SpreadsheetWrite;
