+++ app/src/domain/sheets.rs
@@ -1,7 +1,5 @@
 pub mod a1_notation;
 pub mod cell_position;
-pub mod cell_range;
 pub mod column;
-pub mod flatten_double_vec;
 pub mod ranges;
 pub mod row;
