+++ app/src/infrastructure/sheets.rs
@@ -1,4 +1,6 @@
 pub mod auth;
+pub mod cell_range;
+pub mod flatten_double_vec;
 pub mod http_client;
 pub mod spreadsheet_manager;
 pub mod spreadsheet_read;
